+++ Assets/GameScript/HotUpdate/Logic/MainCenter/MainTopEles.cs
@@ -24,7 +24,7 @@ namespace MainCenter
             CurrencyListCom currencyListCom = (CurrencyListCom)_currencyListCom;
             currencyListCom.SetData(mCurrencyIds); //CurrencyListCom.cs
 
-            EventCenter.Instance.Bind<string>(EventEnum.EE_test, OnEventTest);
+            EventCenter.Instance.Bind<string>((int)EventEnum.EE_test, OnEventTest);
         }
 
         private void OnClickItemFuncList(int index, GObject obj)
@@ -56,7 +56,7 @@ namespace MainCenter
         public override void Dispose()
         {
             base.Dispose();
-            EventCenter.Instance.UnBind<string>(EventEnum.EE_test, OnEventTest);
+            EventCenter.Instance.UnBind<string>((int)EventEnum.EE_test, OnEventTest);
         }
     }
 }
